# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blackwell Time API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Holiday": {
        "type": "object",
        "properties": {
          "holidayId": {
            "type": "string",
            "format": "uuid"
          },
          "holidayName": {
            "type": "string",
            "example": "Labor Day"
          },
          "holidayDate": {
            "type": "string",
            "format": "date",
            "example": "2026-09-07"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/holiday/{clientId}/{holidayId}": {
      "put": {
        "operationId": "v1UpdateHoliday",
        "summary": "Update an existing holiday",
        "description": "holidayId is taken from the path — ignored if present in the request body.",
        "tags": [
          "Holiday"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "holidayId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Holiday"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Holiday updated"
          },
          "404": {
            "description": "Client or holiday not found"
          }
        }
      },
      "delete": {
        "operationId": "v1DeleteHoliday",
        "summary": "Delete a holiday",
        "tags": [
          "Holiday"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "holidayId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Holiday deleted"
          },
          "404": {
            "description": "Client or holiday not found"
          }
        }
      }
    },
    "/api/v1/holiday/{clientId}": {
      "post": {
        "operationId": "v1CreateHoliday",
        "summary": "Create a new holiday",
        "description": "holidayId is server-generated — ignored if present in the request body.",
        "tags": [
          "Holiday"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Holiday"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Holiday created"
          },
          "404": {
            "description": "Client not found"
          }
        }
      },
      "get": {
        "operationId": "v1GetHolidays",
        "summary": "Get all holidays for a client",
        "tags": [
          "Holiday"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of holidays",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Holiday"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Client not found"
          }
        }
      }
    },
...
  },
...
}